src/requirements.txt
Django==2.0.4
djangorestframework==3.8.2
pytz==2018.4
Pillow==5.0.0
